+++ flowserv/model/parameter/boolean.py
@@ -13,6 +13,7 @@ additional properties to the base parameter class.
 from flowserv.model.parameter.base import ParameterBase
 
 import flowserv.error as err
+import flowserv.model.parameter.base as pd
 import flowserv.util as util
 
 
@@ -80,21 +81,21 @@ class BoolParameter(ParameterBase):
             try:
                 util.validate_doc(
                     doc,
-                    mandatory=['id', 'type', 'name', 'index', 'isRequired'],
-                    optional=['description', 'defaultValue', 'module']
+                    mandatory=[pd.ID, pd.TYPE, pd.NAME, pd.INDEX, pd.REQUIRED],
+                    optional=[pd.DESC, pd.DEFAULT, pd.MODULE]
                 )
             except ValueError as ex:
                 raise err.InvalidParameterError(str(ex))
-            if doc['type'] != PARA_BOOL:
-                raise ValueError("invalid type '{}'".format(doc['type']))
+            if doc[pd.TYPE] != PARA_BOOL:
+                raise ValueError("invalid type '{}'".format(doc[pd.TYPE]))
         return cls(
-            para_id=doc['id'],
-            name=doc['name'],
-            index=doc['index'],
-            description=doc.get('description'),
-            default_value=doc.get('defaultValue'),
-            is_required=doc['isRequired'],
-            module_id=doc.get('module')
+            para_id=doc[pd.ID],
+            name=doc[pd.NAME],
+            index=doc[pd.INDEX],
+            description=doc.get(pd.DESC),
+            default_value=doc.get(pd.DEFAULT),
+            is_required=doc[pd.REQUIRED],
+            module_id=doc.get(pd.MODULE)
         )
 
     def to_argument(self, value):