+++ flowserv/tests/serialize.py
@@ -9,6 +9,7 @@
 """Helper methods to test object serialization."""
 
 import flowserv.util as util
+import flowserv.model.parameter.base as pd
 import flowserv.model.workflow.state as st
 
 
@@ -111,7 +112,7 @@ def validate_ranking(doc):
     )
     # Schema columns
     for col in doc['schema']:
-        util.validate_doc(doc=col, mandatory=['id', 'name', 'type'])
+        util.validate_doc(doc=col, mandatory=['id', 'name', 'dtype'])
     # Run results
     for entry in doc['ranking']:
         util.validate_doc(doc=entry, mandatory=['run', 'group', 'results'])
@@ -177,7 +178,11 @@ def validate_run_handle(doc, state):
     assert doc['state'] == state
     if state == st.STATE_SUCCESS:
         for r in doc['files']:
-            util.validate_doc(doc=r, mandatory=['id', 'name'])
+            util.validate_doc(
+                doc=r,
+                mandatory=['id', 'name'],
+                optional=['title', 'caption', 'mimeType', 'widget', 'format']
+            )
 
 
 def validate_run_listing(doc):
@@ -291,8 +296,8 @@ def validate_parameter(doc):
     """
     util.validate_doc(
         doc=doc,
-        mandatory=['id', 'name', 'type', 'index', 'isRequired'],
-        optional=['description', 'defaultValue', 'target', 'values', 'module']
+        mandatory=[pd.ID, pd.TYPE, pd.NAME, pd.INDEX, pd.REQUIRED],
+        optional=[pd.DESC, pd.DEFAULT, pd.MODULE, 'target', 'values', 'range']
     )